# PERF: remove use of Panel & perf in rolling corr/cov