closes #17917

```
    before     after       ratio
  [aa9e0024] [872fe711]
-     1.70s    16.02ms      0.01  rolling.Pairwise.time_pairwise(None, 'corr', True)
-     1.84s    17.12ms      0.01  rolling.Pairwise.time_pairwise(10, 'corr', True)
-     1.84s    16.98ms      0.01  rolling.Pairwise.time_pairwise(1000, 'corr', True)
-     1.74s    15.59ms      0.01  rolling.Pairwise.time_pairwise(None, 'cov', True)
-     1.77s    15.63ms      0.01  rolling.Pairwise.time_pairwise(1000, 'cov', True)
-     1.83s    14.62ms      0.01  rolling.Pairwise.time_pairwise(10, 'cov', True)
```