@jreback 2018-01-15
cc @grumpyquant

if you'd give this a test drive. handling the MI is somewhat non-trivial here.